Coerce type wildcard in maps to string for json.
The yaml libraries we're starting with emits interfaces for keys, but golang stdlib json doesn't accept interfaces as keys.  In practice they're all strings anyway.  So far.

Much more of the validation suite is now passing, with most of the failures being on things that we should be rejecting now in ygp/yamelot.