New test format for many cases per file; implement for Haskell
In the Haskell implementation we've been writing a bunch of small test
cases into its own unit tests, because the integration test format
we've been using is kind of cumbersome.  Make a new format where
 * we can have a bunch of small test cases in one file, so the overhead
   is lower for writing a bunch of them;
 * we can have comments;
 * we can test behavior on documents that lack a final newline;
 * we can mark some tests as expected to fail in some implementations.

I'm not totally happy with the xfail markings, because they really
belong more to an implementation than to the tests themselves, which
are shared.  But this is the easiest place to put them at least for now.

Should document the format, the better to implement elsewhere.

Also in copying a bunch of literal-scalar-related tests from the
Haskell unit tests to here, I find that we're wrong on a couple of
them -- we parse them, as we should, but we get the wrong result.
Good thing this format goes beyond "does it parse"!  The Haskell
unit tests didn't, because that was going to be more of a pain.